move `generateSDKTitle` into 'shared'
