+++ Composer/packages/extensions/visual-designer/src/components/nodes/utils.ts
@@ -6,6 +6,7 @@ import { ConceptLabels } from '@bfc/shared';
 
 import { ObiTypes } from '../../constants/ObiTypes';
 
+// TODO: [schema-driven] retire this helper after migrating to uischema
 export function getFriendlyName(data) {
   // use the developer-specified name if it exists...
   if (get(data, '$designer.name')) {
@@ -30,18 +31,4 @@ export function getUserAnswersTitle($type: string): string {
   }
 
   return '';
-}
-
-const truncateSDKType = $type => (typeof $type === 'string' ? $type.split('Microsoft.')[1] : '');
-
-/**
- * Title priority: $designer.name > override title > title from sdk schema > $type suffix
- * @param overrideTitle customized title
- */
-export function generateSDKTitle(data, overrideTitle?: string) {
-  const $type = get(data, '$type');
-  const titleFrom$designer = get(data, '$designer.name');
-  const titleFromShared = get(ConceptLabels, [$type, 'title']);
-  const titleFrom$type = truncateSDKType($type);
-  return titleFrom$designer || overrideTitle || titleFromShared || titleFrom$type;
 }
